find byte + memmove fix + substr done

+++ src/ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: ketrevis <ketrevis@student42.fr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 12:40:08 by ketrevis          #+#    #+#             */
-/*   Updated: 2023/09/09 12:40:09 by ketrevis         ###   ########.fr       */
+/*   Updated: 2023/09/09 14:51:37 by ketrevis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int		i;
+	size_t		i;
 	char	*tmp_src;
 	char	*tmp_dst;
